move header to general component package
